% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifUtils.r
\name{consensusIUPAC}
\alias{consensusIUPAC}
\title{Consensus IUPAC}
\usage{
consensusIUPAC(mstring)
}
\arguments{
\item{mstring}{a base pare string with only nucleotides eg A[AG]AGT}
}
\description{
The reverse of IUPACtoBase, takes base nucleotides and transforms
them them to their IUPAC equivelent
}
\examples{
consenusIUPAC( IUPACtoBase("ARYS"))
# [1] "ARYS"
}

