% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifUtils.r
\name{ispalandrome}
\alias{ispalandrome}
\title{Is Palandrome}
\usage{
ispalandrome(stringIn)
}
\arguments{
\item{stringIn}{a string, limited to IUPAC characters}
}
\value{
TRUE if palandrme FALSE if not
}
\description{
Tests if biostring is a palandrome... takes IUPAC caracters
as input and returns a TRUE or FALSE
}
\examples{
stringIn<-"CANNTG" # NN E-Box
ispalandrome(stringIn)
> TRUE
stringIn<-"GATTAG" 
ispalandrome(stringIn)
> FALSE
}

