% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifUtils.r
\name{motifString}
\alias{motifString}
\title{Motif String}
\usage{
motifString(x)
}
\arguments{
\item{x}{a matrix of values representing AGCT in a motif}
}
\description{
A scoring mechanism to convert PWM to strings
}
\examples{
motifString(matrix(c(1,1,1,1,7,1,1,1),ncol=2))
# [1] "[ACGT]A"
}

