% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifUtils.r
\name{pwm2consensus}
\alias{pwm2consensus}
\title{PWM to Consensus Motif}
\usage{
pwm2consensus(x)
}
\arguments{
\item{x}{A PWM loaded by loadPWMs}
}
\value{
The consensus IUPAC string equivelent
}
\description{
A caller for ConsesnusIUPAC, Recives PWM and returns a consusensus motif
The PWMs should be loaded through the loadPWM utility.
}
\examples{
library(Biostrings)
categories<-do.call(function(x)as.character(unlist(read.table(x))),
as.list("~/Dropbox/UTX-Alex/jan/catagories"))
allData<-readDNAStringSet("~/Dropbox/UTX-Alex/jan/combined.fasta")
heights<-read.table("~/Dropbox/UTX-Alex/jan/combined_heights.bed")
data<-heights[,4:dim(heights)[2]]
pcs<-pca(data)
plotPCs(pcs$eigenVector,c(1,3),pcs$normData,categories)
objEryt<-list(1,"gt",1) # pc=1 function = greater than 1 sd of the mean
objTALL<-list(1,"lt",1) # pc=1 function = less than - 1 sd of the mean
reg<-applyPeakPartitions(pcs$eigenVectors,list(objEryt,objTALL))
#homerWrapper(allData,objEryt,objTALL,
#             "~/Masters/mulcal/inst/lib/homer-4.7/bin/homer2",
#             inst/data/eryt_jurk_1.pwm)
}

